Reduce memory usage of legacy dependency map
